Added tkltest-ui internal option for rted similarity threshold; moved ui config options to util.ui package
